Allow extraction of long path names.
Use the extended path name syntax when extracting files, so that files
can be extracted even when the resulting file name exceeds the
standard Windows 260 character limit.

The .NET libraries do not permit the use of extended path names, so
some operations (setting the file time-stamps and attributes) have to
be done by Win32 calls.